requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.41
psycopg2-binary==2.9.9
paho-mqtt==1.6.1
gunicorn==22.0.0
alembic==1.13.2
python-dotenv==1.0.1
Werkzeug==3.1.3
